sidenav+segmentedcontrol
